src/main/java/com/scaler/stocktradingapp/service/UserService.java
package com.scaler.stocktradingapp.service;

import org.springframework.stereotype.Service;

import com.scaler.stocktradingapp.model.User;
import com.scaler.stocktradingapp.repository.UserRepository;
import org.springframework.stereotype.Service;

import java.util.Optional;

@Service
public class UserService {
    private final UserRepository userRepository;

    public UserService(UserRepository userRepository) {
        this.userRepository = userRepository;
    }

    public User createUser(User user) {
        return userRepository.save(user);
    }

    public Optional<User> getUserById(Long userId) {
        return userRepository.findById(userId);
    }
}